feat: comprehensive database upgrade
- Add organizational_units table (12 departments)
- Add approval_workflows and approval_requests tables
- Add official_correspondence table (مکاتبات اداری)
- Add staff_transfers table (جابجایی نیرو)
- Add medical_equipment table
- Add sim_cards table
- Convert ALL 147 enum columns to string (MySQL best practice)
- Fix financial decimal fields (rial → unsignedBigInteger)
- Add missing indexes on all tables
- Fix Vehicle.php (remove non-existent assigned_driver_id)
- Fix FormTemplate.php (remove duplicate created_by)
- Fix FormSubmission.php (add missing HasMany import)
- Add Iranian health system research data
- Add comprehensive organizational structure documentation

+++ database/migrations/2026_07_30_000024_create_vehicle_trips_table.php
@@ -26,7 +26,7 @@ return new class extends Migration
             $table->integer('end_mileage')->nullable();
             $table->integer('total_distance')->nullable();
             $table->decimal('fuel_filled', 6, 1)->nullable();
-            $table->decimal('fuel_cost', 12, 2)->nullable();
+            $table->unsignedBigInteger('fuel_cost')->nullable();
             $table->text('passenger_list')->nullable();
             $table->text('trip_purpose');
             $table->text('notes')->nullable();
